model-1/requirements.txt
torch
torchvision
Pillow
numpy
matplotlib
seaborn
scikit-learn
